+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+
+0.6.1
+
+* Add buildTemplate() to convert a Token back to a string
+
+0.6.0
+
+* Change syntax of booleans and numbers to no longer be cast if written as
+  strings e.g on=yes is a boolean of value true but on="yes" is a string of
+  value "yes".
+
 0.0.1
 
 The first release